- Changed buffer index type to size_t

+++ include/doremir/buffer.h
@@ -16,8 +16,8 @@ doremir_buffer_t doremir_buffer_copy(doremir_buffer_t);
 doremir_buffer_t doremir_buffer_resize(size_t, doremir_buffer_t);
 void doremir_buffer_destroy(doremir_buffer_t);
 size_t doremir_buffer_size(doremir_buffer_t);
-uint8_t doremir_buffer_peek(doremir_buffer_t, int);
-void doremir_buffer_poke(doremir_buffer_t, int, uint8_t);
+uint8_t doremir_buffer_peek(doremir_buffer_t, size_t);
+void doremir_buffer_poke(doremir_buffer_t, size_t, uint8_t);
 
 /** @}
     @}
